# faster viterbi decoder