在Viterbi decode过程中，使用动态内存会降低解码效率。故采用连续内存的解码表以提高效率。

benchmark结果如下，使用20*20的解码表，无constrain，10万轮迭代

|      | old decoder | new decoder      |
|------|-------------|------------------|
| time | 73.78s      | 11.61s (-84.26%) |


序列标注任务实验结果如下，其中train的时间为一轮迭代学习所用的时间（包括了feature extract, decode, param update）


| task | dataset (#instances) | old decoder | new decoder     |
|-----|-------------------|-------------|-----------------|
| cws | train (292,541)     | 217s        | 205s (-5.53%)   |
|     | test (7,500)        | 3.64s       | 3.38s (-7.14%)  |
| pos | train (64,654)      | 190s        | 42s (-77.89%)   |
|     | test (3,144)        | 9.94s       | 1.81s (-81.79%) |
| ner | train (34,426)      | 31s         | 22s (-29.03%)   |
|     | test (10,000)       | 5.79s       | 3.68s (-36.44)  |

可以看出，解码任务的constrain越少，则所有合法的状态序列的个数越多，消除动态内存申请所带来的效率提升越大。